Still a bit of work

+++ Passengers/src/main/java/be/vinci/chattycar/passengers/PassengersService.java
@@ -3,8 +3,8 @@ package be.vinci.chattycar.passengers;
 import be.vinci.chattycar.passengers.data.PassengersRepository;
 import be.vinci.chattycar.passengers.data.TripsProxy;
 import be.vinci.chattycar.passengers.data.UsersProxy;
-import be.vinci.chattycar.passengers.exceptions.InscriptionExists400Exception;
-import be.vinci.chattycar.passengers.exceptions.InscriptionNotFound404Exception;
+import be.vinci.chattycar.passengers.exceptions.PassengerExists400Exception;
+import be.vinci.chattycar.passengers.exceptions.PassengerNotFound404Exception;
 import be.vinci.chattycar.passengers.exceptions.InscriptionNotValidException;
 import be.vinci.chattycar.passengers.exceptions.NoSeatLeft400Exception;
 import be.vinci.chattycar.passengers.exceptions.TripNotFound404Exception;
@@ -13,7 +13,6 @@ import be.vinci.chattycar.passengers.exceptions.UserNotFound404Exception;
 import be.vinci.chattycar.passengers.models.Passenger;
 import be.vinci.chattycar.passengers.models.PassengerTrips;
 import be.vinci.chattycar.passengers.models.Passengers;
-import be.vinci.chattycar.passengers.models.Trip;
 import java.util.Optional;
 
 public class PassengersService {
@@ -35,7 +34,7 @@ public class PassengersService {
    * @throws TripNotFound404Exception
    */
   public Passengers getPassengers(Integer tripId) throws TripNotFound404Exception {
-    if (!repository.existsById(tripId)) throw new TripNotFound404Exception();
+    if (!repository.existsByTripId(tripId)) throw new TripNotFound404Exception();
     return repository.findAllPassengersByTripId(tripId);
   }
 
@@ -45,7 +44,7 @@ public class PassengersService {
    * @return True if the passenger of a trip were deleted, false if it couldn't be found
    */
   public boolean deleteOne(Integer tripId) throws TripNotFound404Exception {
-    if (!repository.existsById(tripId)) throw new TripNotFound404Exception();
+    if (!repository.existsByTripId(tripId)) throw new TripNotFound404Exception();
     repository.deleteByTripId(tripId);
     return true;
   }
@@ -57,26 +56,25 @@ public class PassengersService {
    * @return The passenger created with its id, or null if it already existed
    */
   public Passenger createOne(Integer tripId, Integer userId) throws TripNotFound404Exception,
-      UserNotFound404Exception, NoSeatLeft400Exception, InscriptionExists400Exception, InternalError {
-    if (repository.existsByIds(tripId, userId)) throw new InscriptionExists400Exception();
-    if(!repository.userExists(userId)) throw new UserNotFound404Exception();
-    if(!repository.tripExists(tripId)) throw new TripNotFound404Exception();
-    if(repository.getTrip(tripId).getAvailableSeating() < 1) throw new NoSeatLeft400Exception();
-    // TODO: 28/11/2022 Comment créer un trip ? 
-    //return repository.save(new Passenger(tripId, userId, "pending"));
-    return null;
+      UserNotFound404Exception, NoSeatLeft400Exception, PassengerExists400Exception, InternalError {
+    if (repository.existsByUserIdAndTripId(tripId, userId)) throw new PassengerExists400Exception();
+    if(!repository.existsByUserId(userId)) throw new UserNotFound404Exception();
+    if(!repository.existsByTripId(tripId)) throw new TripNotFound404Exception();
+    if(repository.getTrip(tripId).getAvailable_seating() < 1) throw new NoSeatLeft400Exception();
+    return repository.save(new Passenger(-1, tripId, userId, "pending"));
   }
 
   /**
-   * Reads a passenger in a trip
+   * Get a passenger status in a trip
    * @param tripId id of the trip
    * @param userId id of the user
-   * @return The passenger, or null if it couldn't be found
+   * @return The passenger status
+   * @throws TripOrUserNotFound404Exception
    */
-  public Passenger readOne(long tripId, long userId) {
-    Optional<Passenger> p = repository.findByIds(tripId, userId);
+  public String getPassengerStatus(long tripId, long userId) throws TripOrUserNotFound404Exception{
+    Optional<Passenger> p = repository.findByUserIdAndTripId(tripId, userId);
     if (!p.isPresent()) throw new TripOrUserNotFound404Exception();
-    return p.get();
+    return p.get().getStatus();
   }
 
   /**
@@ -84,36 +82,41 @@ public class PassengersService {
    * @param tripId id of the trip
    * @param userId id of the user
    * @return True if the inscription was updated, or null if it couldn't be found
+   * @throws InscriptionNotValidException
+   * @throws UserNotFound404Exception
+   * @throws TripNotFound404Exception
    */
   public boolean updateOne(Integer tripId, Integer userId, String status)
       throws InscriptionNotValidException, UserNotFound404Exception, TripNotFound404Exception{
-    if(!repository.userExists(userId)) throw new UserNotFound404Exception();
-    if(!repository.tripExists(tripId)) throw new TripNotFound404Exception();
-    if(!repository.existsByIds(tripId, userId)
-        || !repository.findByIds(tripId, userId).get().getStatus().equals("pending")
+    if(!repository.existsByUserId(userId)) throw new UserNotFound404Exception();
+    if(!repository.existsByTripId(tripId)) throw new TripNotFound404Exception();
+    if(!repository.existsByUserIdAndTripId(tripId, userId)
+        || !repository.findByUserIdAndTripId(tripId, userId).get().getStatus().equals("pending")
         || !(status.equals("accepted") || status.equals("refused"))) throw new InscriptionNotValidException();
-    // TODO: 28/11/2022 Comment créer un new passenger
-    //repository.save(new Passenger(tripId, userId, "pending"));
+    repository.save(new Passenger(-1, tripId, userId, "pending"));
     return true;
   }
 
   /**
-   * Deletes a review from repository
+   * Deletes a passenger from repository
    * @param tripId id of the trip
    * @param userId id of the user
    * @return True if the review was deleted, false if it couldn't be found
+   * @throws PassengerNotFound404Exception
+   * @throws UserNotFound404Exception
+   * @throws TripNotFound404Exception
    */
   public boolean deleteOne(Integer tripId, Integer userId)
-      throws InscriptionNotFound404Exception, UserNotFound404Exception, TripNotFound404Exception {
-    if(!repository.userExists(userId)) throw new UserNotFound404Exception();
-    if(!repository.tripExists(tripId)) throw new TripNotFound404Exception();
-    if(!repository.existsByIds(tripId, userId)) throw new InscriptionNotFound404Exception();
+      throws PassengerNotFound404Exception, UserNotFound404Exception, TripNotFound404Exception {
+    if(!repository.existsByUserId(userId)) throw new UserNotFound404Exception();
+    if(!repository.existsByTripId(tripId)) throw new TripNotFound404Exception();
+    if(!repository.existsByUserIdAndTripId(tripId, userId)) throw new PassengerNotFound404Exception();
     repository.deleteByIds(tripId, userId);
     return true;
   }
 
   /**
-   * Reads all reviews from a user
+   * Reads all trips from a passenger
    * @param userId id of the user
    * @return The list of trips from this user
    */
@@ -125,7 +128,8 @@ public class PassengersService {
    * Deletes all trips from a user
    * @param userId id of the user
    */
-  public void deleteTripsFromPassenger(Integer userId) {
+  public void deleteTripsFromPassenger(Integer userId) throws UserNotFound404Exception {
+    if(!repository.existsByUserId(userId)) throw new UserNotFound404Exception();
     repository.deleteByUserId(userId);
   }
